@@ -13,7 +13,7 @@ ext {
     siteUrl = 'https://github.com/ChathuraHettiarachchi/TimeAgo'
     gitUrl = 'https://github.com/ChathuraHettiarachchi/TimeAgo.git'
 
-    libraryVersion = '1.1.8'
+    libraryVersion = '1.2'
 
     developerId = 'ChathuraHettiarachchi'
     developerName = 'Chathura Hettiarachchi'
@@ -25,14 +25,13 @@ ext {
 }
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.3"
+    compileSdkVersion 28
 
     defaultConfig {
-        minSdkVersion 9
-        targetSdkVersion 23
+        minSdkVersion 21
+        targetSdkVersion 28
         versionCode 1
-        versionName "1.0"
+        versionName "1.2"
     }
     buildTypes {
         release {
@@ -43,9 +42,9 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    testCompile 'junit:junit:4.12'
-    compile 'com.android.support:appcompat-v7:23.4.0'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    testImplementation 'junit:junit:4.12'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
 }
 
 // Place it at the end of the file
