@@ -1,7 +1,7 @@
 package com.choota.dev.timeago;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.widget.TextView;
 
 import com.choota.dev.ctimeago.TimeAgo;
